Fixed formatting of generated code
